Initial implementation of wrapping of decor_t
Wrap decor_t with namedtuple or structseq (choose by
#define USE_PY_STRUCTSEQ , due to bug in python structseq will cause
crashes).

Wrap all members of decor_t except for structures that can't be wrapped by
pygobject nor pycairo, and except for the draw function pointer, which we
will not wrap.

Pass the wrapped decor_t to the python side and print it to stdout just to
see if it works.